feat: add `utils` file
