+++ src/graphql/helpers/post.js
@@ -1,5 +1,3 @@
-import rp from 'request-promise-native';
-
 const getFeedParams = ({ user, post }, args, rankBy, filters = {}) => {
   const userId = user ? user.userId : null;
 
@@ -13,32 +11,9 @@ const getFeedParams = ({ user, post }, args, rankBy, filters = {}) => {
   };
 };
 
-const splitArrayStr = str => (str ? str.split(',') : null);
-
 const assignType = type => x => Object.assign({}, x, { type });
-
-const fetchToiletAd = async (ip, config) => {
-  try {
-    const res = await rp({
-      url: `${config.monetizationUrl}/a/toilet`,
-      method: 'GET',
-      headers: {
-        'x-forwarded-for': ip,
-      },
-    });
-
-    return JSON.parse(res);
-  } catch (err) {
-    // eslint-disable-next-line
-    console.warn('failed to fetch ad from monetization service', { err });
-
-    return [];
-  }
-};
 
 export default {
   getFeedParams,
-  splitArrayStr,
   assignType,
-  fetchToiletAd,
 };
